+++ gcp-auth-extension/src/test/java/io/opentelemetry/contrib/gcp/auth/GcpAuthAutoConfigurationCustomizerProviderTest.java
@@ -8,6 +8,7 @@ package io.opentelemetry.contrib.gcp.auth;
 import static io.opentelemetry.contrib.gcp.auth.GcpAuthAutoConfigurationCustomizerProvider.GCP_USER_PROJECT_ID_KEY;
 import static io.opentelemetry.contrib.gcp.auth.GcpAuthAutoConfigurationCustomizerProvider.QUOTA_USER_PROJECT_HEADER;
 import static io.opentelemetry.contrib.gcp.auth.GcpAuthAutoConfigurationCustomizerProvider.SIGNAL_TYPE_ALL;
+import static io.opentelemetry.contrib.gcp.auth.GcpAuthAutoConfigurationCustomizerProvider.SIGNAL_TYPE_LOGS;
 import static io.opentelemetry.contrib.gcp.auth.GcpAuthAutoConfigurationCustomizerProvider.SIGNAL_TYPE_METRICS;
 import static io.opentelemetry.contrib.gcp.auth.GcpAuthAutoConfigurationCustomizerProvider.SIGNAL_TYPE_TRACES;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -29,10 +30,14 @@ import io.opentelemetry.api.metrics.LongCounter;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.common.ComponentLoader;
 import io.opentelemetry.context.Scope;
+import io.opentelemetry.exporter.otlp.http.logs.OtlpHttpLogRecordExporter;
+import io.opentelemetry.exporter.otlp.http.logs.OtlpHttpLogRecordExporterBuilder;
 import io.opentelemetry.exporter.otlp.http.metrics.OtlpHttpMetricExporter;
 import io.opentelemetry.exporter.otlp.http.metrics.OtlpHttpMetricExporterBuilder;
 import io.opentelemetry.exporter.otlp.http.trace.OtlpHttpSpanExporter;
 import io.opentelemetry.exporter.otlp.http.trace.OtlpHttpSpanExporterBuilder;
+import io.opentelemetry.exporter.otlp.logs.OtlpGrpcLogRecordExporter;
+import io.opentelemetry.exporter.otlp.logs.OtlpGrpcLogRecordExporterBuilder;
 import io.opentelemetry.exporter.otlp.metrics.OtlpGrpcMetricExporter;
 import io.opentelemetry.exporter.otlp.metrics.OtlpGrpcMetricExporterBuilder;
 import io.opentelemetry.exporter.otlp.trace.OtlpGrpcSpanExporter;
@@ -43,10 +48,13 @@ import io.opentelemetry.sdk.autoconfigure.AutoConfiguredOpenTelemetrySdkBuilder;
 import io.opentelemetry.sdk.autoconfigure.internal.SpiHelper;
 import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
 import io.opentelemetry.sdk.autoconfigure.spi.ConfigurationException;
+import io.opentelemetry.sdk.autoconfigure.spi.logs.ConfigurableLogRecordExporterProvider;
 import io.opentelemetry.sdk.autoconfigure.spi.metrics.ConfigurableMetricExporterProvider;
 import io.opentelemetry.sdk.autoconfigure.spi.traces.ConfigurableSpanExporterProvider;
 import io.opentelemetry.sdk.common.CompletableResultCode;
 import io.opentelemetry.sdk.common.export.MemoryMode;
+import io.opentelemetry.sdk.logs.data.LogRecordData;
+import io.opentelemetry.sdk.logs.export.LogRecordExporter;
 import io.opentelemetry.sdk.metrics.Aggregation;
 import io.opentelemetry.sdk.metrics.InstrumentType;
 import io.opentelemetry.sdk.metrics.data.AggregationTemporality;
@@ -101,6 +109,7 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
 
   @Captor private ArgumentCaptor<Supplier<Map<String, String>>> traceHeaderSupplierCaptor;
   @Captor private ArgumentCaptor<Supplier<Map<String, String>>> metricHeaderSupplierCaptor;
+  @Captor private ArgumentCaptor<Supplier<Map<String, String>>> logRecordHeaderSupplierCaptor;
 
   private static final ImmutableMap<String, String> defaultOtelPropertiesSpanExporter =
       ImmutableMap.of(
@@ -125,6 +134,19 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
           "otlp",
           "otel.logs.exporter",
           "none",
+          "otel.resource.attributes",
+          "foo=bar");
+
+  private static final ImmutableMap<String, String> defaultOtelPropertiesLogRecordExporter =
+      ImmutableMap.of(
+          "otel.exporter.otlp.logs.endpoint",
+          "https://telemetry.googleapis.com/v1/logs",
+          "otel.traces.exporter",
+          "none",
+          "otel.metrics.exporter",
+          "none",
+          "otel.logs.exporter",
+          "otlp",
           "otel.resource.attributes",
           "foo=bar");
 
@@ -372,6 +394,121 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
     }
   }
 
+  // TODO: Use parameterized test for testing logs customizer for http & grpc.
+  @Test
+  void testLogRecordCustomizerOtlpHttp() {
+    // Set resource project system property
+    System.setProperty(
+        ConfigurableOption.GOOGLE_CLOUD_PROJECT.getSystemProperty(), DUMMY_GCP_RESOURCE_PROJECT_ID);
+    System.setProperty(
+        ConfigurableOption.GOOGLE_OTEL_AUTH_TARGET_SIGNALS.getSystemProperty(), SIGNAL_TYPE_LOGS);
+    // Prepare mocks
+    prepareMockBehaviorForGoogleCredentials();
+    OtlpHttpLogRecordExporter mockOtlpHttpLogRecordExporter =
+        Mockito.mock(OtlpHttpLogRecordExporter.class);
+    OtlpHttpLogRecordExporterBuilder otlpLogRecordExporterBuilder =
+        OtlpHttpLogRecordExporter.builder();
+    OtlpHttpLogRecordExporterBuilder spyOtlpHttpLogRecordExporterBuilder =
+        Mockito.spy(otlpLogRecordExporterBuilder);
+    List<LogRecordData> exportedLogRecords = new ArrayList<>();
+    configureHttpMockLogRecordExporter(
+        mockOtlpHttpLogRecordExporter, spyOtlpHttpLogRecordExporterBuilder, exportedLogRecords);
+
+    // begin assertions
+    try (MockedStatic<GoogleCredentials> googleCredentialsMockedStatic =
+        Mockito.mockStatic(GoogleCredentials.class)) {
+      googleCredentialsMockedStatic
+          .when(GoogleCredentials::getApplicationDefault)
+          .thenReturn(mockedGoogleCredentials);
+
+      OpenTelemetrySdk sdk = buildOpenTelemetrySdkWithExporter(mockOtlpHttpLogRecordExporter);
+      generateTestLogRecord(sdk);
+      CompletableResultCode code = sdk.shutdown();
+      CompletableResultCode joinResult = code.join(10, TimeUnit.SECONDS);
+      assertThat(joinResult.isSuccess()).isTrue();
+
+      Mockito.verify(mockOtlpHttpLogRecordExporter, Mockito.times(1)).toBuilder();
+      Mockito.verify(spyOtlpHttpLogRecordExporterBuilder, Mockito.times(1))
+          .setHeaders(logRecordHeaderSupplierCaptor.capture());
+      assertThat(logRecordHeaderSupplierCaptor.getValue().get().size()).isEqualTo(2);
+      assertThat(authHeadersQuotaProjectIsPresent(logRecordHeaderSupplierCaptor.getValue().get()))
+          .isTrue();
+
+      Mockito.verify(mockOtlpHttpLogRecordExporter, Mockito.atLeast(1))
+          .export(Mockito.anyCollection());
+
+      assertThat(exportedLogRecords)
+          .hasSizeGreaterThan(0)
+          .allSatisfy(
+              logRecordData -> {
+                assertThat(logRecordData.getResource().getAttributes().asMap())
+                    .containsEntry(
+                        AttributeKey.stringKey(GCP_USER_PROJECT_ID_KEY),
+                        DUMMY_GCP_RESOURCE_PROJECT_ID)
+                    .containsEntry(AttributeKey.stringKey("foo"), "bar");
+                assertThat(logRecordData.getAttributes().asMap())
+                    .containsKey(AttributeKey.longKey("work_loop"));
+              });
+    }
+  }
+
+  @Test
+  void testLogRecordCustomizerOtlpGrpc() {
+    // Set resource project system property
+    System.setProperty(
+        ConfigurableOption.GOOGLE_CLOUD_PROJECT.getSystemProperty(), DUMMY_GCP_RESOURCE_PROJECT_ID);
+    System.setProperty(
+        ConfigurableOption.GOOGLE_OTEL_AUTH_TARGET_SIGNALS.getSystemProperty(), SIGNAL_TYPE_LOGS);
+    // Prepare mocks
+    prepareMockBehaviorForGoogleCredentials();
+    OtlpGrpcLogRecordExporter mockOtlpGrpcLogRecordExporter =
+        Mockito.mock(OtlpGrpcLogRecordExporter.class);
+    OtlpGrpcLogRecordExporterBuilder otlpLogRecordExporterBuilder =
+        OtlpGrpcLogRecordExporter.builder();
+    OtlpGrpcLogRecordExporterBuilder spyOtlpGrpcLogRecordExporterBuilder =
+        Mockito.spy(otlpLogRecordExporterBuilder);
+    List<LogRecordData> exportedLogRecords = new ArrayList<>();
+    configureGrpcMockLogRecordExporter(
+        mockOtlpGrpcLogRecordExporter, spyOtlpGrpcLogRecordExporterBuilder, exportedLogRecords);
+
+    // begin assertions
+    try (MockedStatic<GoogleCredentials> googleCredentialsMockedStatic =
+        Mockito.mockStatic(GoogleCredentials.class)) {
+      googleCredentialsMockedStatic
+          .when(GoogleCredentials::getApplicationDefault)
+          .thenReturn(mockedGoogleCredentials);
+
+      OpenTelemetrySdk sdk = buildOpenTelemetrySdkWithExporter(mockOtlpGrpcLogRecordExporter);
+      generateTestLogRecord(sdk);
+      CompletableResultCode code = sdk.shutdown();
+      CompletableResultCode joinResult = code.join(10, TimeUnit.SECONDS);
+      assertThat(joinResult.isSuccess()).isTrue();
+
+      Mockito.verify(mockOtlpGrpcLogRecordExporter, Mockito.times(1)).toBuilder();
+      Mockito.verify(spyOtlpGrpcLogRecordExporterBuilder, Mockito.times(1))
+          .setHeaders(logRecordHeaderSupplierCaptor.capture());
+      assertThat(logRecordHeaderSupplierCaptor.getValue().get().size()).isEqualTo(2);
+      assertThat(authHeadersQuotaProjectIsPresent(logRecordHeaderSupplierCaptor.getValue().get()))
+          .isTrue();
+
+      Mockito.verify(mockOtlpGrpcLogRecordExporter, Mockito.atLeast(1))
+          .export(Mockito.anyCollection());
+
+      assertThat(exportedLogRecords)
+          .hasSizeGreaterThan(0)
+          .allSatisfy(
+              logRecordData -> {
+                assertThat(logRecordData.getResource().getAttributes().asMap())
+                    .containsEntry(
+                        AttributeKey.stringKey(GCP_USER_PROJECT_ID_KEY),
+                        DUMMY_GCP_RESOURCE_PROJECT_ID)
+                    .containsEntry(AttributeKey.stringKey("foo"), "bar");
+                assertThat(logRecordData.getAttributes().asMap())
+                    .containsKey(AttributeKey.longKey("work_loop"));
+              });
+    }
+  }
+
   @Test
   void testCustomizerFailWithMissingResourceProject() {
     System.setProperty(
@@ -572,6 +709,17 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
     configureGrpcMockMetricExporter(
         mockOtlpGrpcMetricExporter, spyOtlpGrpcMetricExporterBuilder, exportedMetrics);
 
+    // Prepare mocked log record exporter
+    OtlpGrpcLogRecordExporter mockOtlpGrpcLogRecordExporter =
+        Mockito.mock(OtlpGrpcLogRecordExporter.class);
+    OtlpGrpcLogRecordExporterBuilder otlpLogRecordExporterBuilder =
+        OtlpGrpcLogRecordExporter.builder();
+    OtlpGrpcLogRecordExporterBuilder spyOtlpGrpcLogRecordExporterBuilder =
+        Mockito.spy(otlpLogRecordExporterBuilder);
+    List<LogRecordData> exportedLogRecords = new ArrayList<>();
+    configureGrpcMockLogRecordExporter(
+        mockOtlpGrpcLogRecordExporter, spyOtlpGrpcLogRecordExporterBuilder, exportedLogRecords);
+
     // configure environment according to test case
     System.setProperty(
         ConfigurableOption.GOOGLE_OTEL_AUTH_TARGET_SIGNALS.getSystemProperty(),
@@ -588,9 +736,11 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
           buildOpenTelemetrySdkWithExporter(
               mockOtlpGrpcSpanExporter,
               mockOtlpGrpcMetricExporter,
+              mockOtlpGrpcLogRecordExporter,
               testCase.getUserSpecifiedOtelProperties());
       generateTestMetric(sdk);
       generateTestSpan(sdk);
+      generateTestLogRecord(sdk);
       CompletableResultCode code = sdk.shutdown();
       CompletableResultCode joinResult = code.join(10, TimeUnit.SECONDS);
       assertThat(joinResult.isSuccess()).isTrue();
@@ -622,6 +772,20 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
         // should be made
         Mockito.verifyNoInteractions(spyOtlpGrpcMetricExporterBuilder);
       }
+
+      // Check Logs modification conditions
+      if (testCase.getExpectedIsLogsSignalModified()) {
+        // If logs signal is expected to be modified, auth headers must be present
+        Mockito.verify(spyOtlpGrpcLogRecordExporterBuilder, Mockito.times(1))
+            .setHeaders(logRecordHeaderSupplierCaptor.capture());
+        assertThat(logRecordHeaderSupplierCaptor.getValue().get().size()).isEqualTo(2);
+        assertThat(authHeadersQuotaProjectIsPresent(logRecordHeaderSupplierCaptor.getValue().get()))
+            .isTrue();
+      } else {
+        // If logs signals is not expected to be modified then no interaction with the builder
+        // should be made
+        Mockito.verifyNoInteractions(spyOtlpGrpcLogRecordExporterBuilder);
+      }
     }
   }
 
@@ -634,6 +798,7 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
                 .setUserSpecifiedOtelProperties(defaultOtelPropertiesSpanExporter)
                 .setExpectedIsMetricsSignalModified(false)
                 .setExpectedIsTraceSignalModified(true)
+                .setExpectedIsLogsSignalModified(false)
                 .build()),
         Arguments.of(
             TargetSignalBehavior.builder()
@@ -641,6 +806,15 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
                 .setUserSpecifiedOtelProperties(defaultOtelPropertiesMetricExporter)
                 .setExpectedIsMetricsSignalModified(true)
                 .setExpectedIsTraceSignalModified(false)
+                .setExpectedIsLogsSignalModified(false)
+                .build()),
+        Arguments.of(
+            TargetSignalBehavior.builder()
+                .setConfiguredTargetSignals("logs")
+                .setUserSpecifiedOtelProperties(defaultOtelPropertiesLogRecordExporter)
+                .setExpectedIsMetricsSignalModified(false)
+                .setExpectedIsTraceSignalModified(false)
+                .setExpectedIsLogsSignalModified(true)
                 .build()),
         Arguments.of(
             TargetSignalBehavior.builder()
@@ -651,14 +825,17 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
                         "https://localhost:4813/v1/metrics",
                         "otel.exporter.otlp.traces.endpoint",
                         "https://localhost:4813/v1/traces",
+                        "otel.exporter.otlp.logs.endpoint",
+                        "https://localhost:4813/v1/logs",
                         "otel.traces.exporter",
                         "otlp",
                         "otel.metrics.exporter",
                         "otlp",
                         "otel.logs.exporter",
-                        "none"))
+                        "otlp"))
                 .setExpectedIsMetricsSignalModified(true)
                 .setExpectedIsTraceSignalModified(true)
+                .setExpectedIsLogsSignalModified(true)
                 .build()),
         Arguments.of(
             TargetSignalBehavior.builder()
@@ -669,14 +846,38 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
                         "https://localhost:4813/v1/metrics",
                         "otel.exporter.otlp.traces.endpoint",
                         "https://localhost:4813/v1/traces",
+                        "otel.exporter.otlp.logs.endpoint",
+                        "https://localhost:4813/v1/logs",
                         "otel.traces.exporter",
                         "otlp",
                         "otel.metrics.exporter",
                         "otlp",
                         "otel.logs.exporter",
-                        "none"))
+                        "otlp"))
                 .setExpectedIsMetricsSignalModified(true)
                 .setExpectedIsTraceSignalModified(true)
+                .setExpectedIsLogsSignalModified(false)
+                .build()),
+        Arguments.of(
+            TargetSignalBehavior.builder()
+                .setConfiguredTargetSignals("logs, traces")
+                .setUserSpecifiedOtelProperties(
+                    ImmutableMap.of(
+                        "otel.exporter.otlp.metrics.endpoint",
+                        "https://localhost:4813/v1/metrics",
+                        "otel.exporter.otlp.traces.endpoint",
+                        "https://localhost:4813/v1/traces",
+                        "otel.exporter.otlp.logs.endpoint",
+                        "https://localhost:4813/v1/logs",
+                        "otel.traces.exporter",
+                        "otlp",
+                        "otel.metrics.exporter",
+                        "otlp",
+                        "otel.logs.exporter",
+                        "otlp"))
+                .setExpectedIsMetricsSignalModified(false)
+                .setExpectedIsTraceSignalModified(true)
+                .setExpectedIsLogsSignalModified(true)
                 .build()),
         Arguments.of(
             TargetSignalBehavior.builder()
@@ -687,14 +888,17 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
                         "https://localhost:4813/v1/metrics",
                         "otel.exporter.otlp.traces.endpoint",
                         "https://localhost:4813/v1/traces",
+                        "otel.exporter.otlp.logs.endpoint",
+                        "https://localhost:4813/v1/logs",
                         "otel.traces.exporter",
                         "otlp",
                         "otel.metrics.exporter",
                         "otlp",
                         "otel.logs.exporter",
-                        "none"))
+                        "otlp"))
                 .setExpectedIsMetricsSignalModified(true)
                 .setExpectedIsTraceSignalModified(true)
+                .setExpectedIsLogsSignalModified(true)
                 .build()),
         Arguments.of(
             TargetSignalBehavior.builder()
@@ -705,6 +909,8 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
                         "https://localhost:4813/v1/metrics",
                         "otel.exporter.otlp.traces.endpoint",
                         "https://localhost:4813/v1/traces",
+                        "otel.exporter.otlp.logs.endpoint",
+                        "https://localhost:4813/v1/logs",
                         "otel.traces.exporter",
                         "none",
                         "otel.metrics.exporter",
@@ -713,24 +919,28 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
                         "none"))
                 .setExpectedIsMetricsSignalModified(false)
                 .setExpectedIsTraceSignalModified(false)
+                .setExpectedIsLogsSignalModified(false)
                 .build()),
         Arguments.of(
             TargetSignalBehavior.builder()
-                .setConfiguredTargetSignals("metric, trace")
+                .setConfiguredTargetSignals("metric, trace, log")
                 .setUserSpecifiedOtelProperties(
                     ImmutableMap.of(
                         "otel.exporter.otlp.metrics.endpoint",
                         "https://localhost:4813/v1/metrics",
                         "otel.exporter.otlp.traces.endpoint",
                         "https://localhost:4813/v1/traces",
+                        "otel.exporter.otlp.logs.endpoint",
+                        "https://localhost:4813/v1/logs",
                         "otel.traces.exporter",
                         "otlp",
                         "otel.metrics.exporter",
                         "otlp",
                         "otel.logs.exporter",
-                        "none"))
+                        "otlp"))
                 .setExpectedIsMetricsSignalModified(false)
                 .setExpectedIsTraceSignalModified(false)
+                .setExpectedIsLogsSignalModified(false)
                 .build()),
         Arguments.of(
             TargetSignalBehavior.builder()
@@ -741,14 +951,17 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
                         "https://localhost:4813/v1/metrics",
                         "otel.exporter.otlp.traces.endpoint",
                         "https://localhost:4813/v1/traces",
+                        "otel.exporter.otlp.logs.endpoint",
+                        "https://localhost:4813/v1/logs",
                         "otel.traces.exporter",
                         "otlp",
                         "otel.metrics.exporter",
                         "otlp",
                         "otel.logs.exporter",
-                        "none"))
+                        "otlp"))
                 .setExpectedIsMetricsSignalModified(true)
                 .setExpectedIsTraceSignalModified(false)
+                .setExpectedIsLogsSignalModified(false)
                 .build()),
         Arguments.of(
             TargetSignalBehavior.builder()
@@ -759,14 +972,17 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
                         "https://localhost:4813/v1/metrics",
                         "otel.exporter.otlp.traces.endpoint",
                         "https://localhost:4813/v1/traces",
+                        "otel.exporter.otlp.logs.endpoint",
+                        "https://localhost:4813/v1/logs",
                         "otel.traces.exporter",
                         "otlp",
                         "otel.metrics.exporter",
                         "otlp",
                         "otel.logs.exporter",
-                        "none"))
+                        "otlp"))
                 .setExpectedIsMetricsSignalModified(false)
                 .setExpectedIsTraceSignalModified(false)
+                .setExpectedIsLogsSignalModified(false)
                 .build()),
         Arguments.of(
             TargetSignalBehavior.builder()
@@ -777,14 +993,38 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
                         "https://localhost:4813/v1/metrics",
                         "otel.exporter.otlp.traces.endpoint",
                         "https://localhost:4813/v1/traces",
+                        "otel.exporter.otlp.logs.endpoint",
+                        "https://localhost:4813/v1/logs",
                         "otel.traces.exporter",
                         "otlp",
                         "otel.metrics.exporter",
                         "otlp",
                         "otel.logs.exporter",
-                        "none"))
+                        "otlp"))
                 .setExpectedIsMetricsSignalModified(false)
                 .setExpectedIsTraceSignalModified(false)
+                .setExpectedIsLogsSignalModified(false)
+                .build()),
+        Arguments.of(
+            TargetSignalBehavior.builder()
+                .setConfiguredTargetSignals("logs, none")
+                .setUserSpecifiedOtelProperties(
+                    ImmutableMap.of(
+                        "otel.exporter.otlp.metrics.endpoint",
+                        "https://localhost:4813/v1/metrics",
+                        "otel.exporter.otlp.traces.endpoint",
+                        "https://localhost:4813/v1/traces",
+                        "otel.exporter.otlp.logs.endpoint",
+                        "https://localhost:4813/v1/logs",
+                        "otel.traces.exporter",
+                        "otlp",
+                        "otel.metrics.exporter",
+                        "otlp",
+                        "otel.logs.exporter",
+                        "otlp"))
+                .setExpectedIsMetricsSignalModified(false)
+                .setExpectedIsTraceSignalModified(false)
+                .setExpectedIsLogsSignalModified(false)
                 .build()),
         Arguments.of(
             TargetSignalBehavior.builder()
@@ -795,14 +1035,17 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
                         "https://localhost:4813/v1/metrics",
                         "otel.exporter.otlp.traces.endpoint",
                         "https://localhost:4813/v1/traces",
+                        "otel.exporter.otlp.logs.endpoint",
+                        "https://localhost:4813/v1/logs",
                         "otel.traces.exporter",
                         "otlp",
                         "otel.metrics.exporter",
                         "otlp",
                         "otel.logs.exporter",
-                        "none"))
+                        "otlp"))
                 .setExpectedIsMetricsSignalModified(false)
                 .setExpectedIsTraceSignalModified(false)
+                .setExpectedIsLogsSignalModified(false)
                 .build()),
         Arguments.of(
             TargetSignalBehavior.builder()
@@ -813,14 +1056,17 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
                         "https://localhost:4813/v1/metrics",
                         "otel.exporter.otlp.traces.endpoint",
                         "https://localhost:4813/v1/traces",
+                        "otel.exporter.otlp.logs.endpoint",
+                        "https://localhost:4813/v1/logs",
                         "otel.traces.exporter",
                         "otlp",
                         "otel.metrics.exporter",
                         "otlp",
                         "otel.logs.exporter",
-                        "none"))
+                        "otlp"))
                 .setExpectedIsMetricsSignalModified(false)
                 .setExpectedIsTraceSignalModified(false)
+                .setExpectedIsLogsSignalModified(false)
                 .build()));
   }
 
@@ -1064,6 +1310,56 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
         .thenReturn(MemoryMode.IMMUTABLE_DATA);
   }
 
+  // Configure necessary behavior on the http mock log record exporters to work.
+  private static void configureHttpMockLogRecordExporter(
+      OtlpHttpLogRecordExporter mockOtlpHttpLogRecordExporter,
+      OtlpHttpLogRecordExporterBuilder spyOtlpHttpLogRecordExporterBuilder,
+      List<LogRecordData> exportedLogRecordContainer) {
+    Mockito.when(spyOtlpHttpLogRecordExporterBuilder.build())
+        .thenReturn(mockOtlpHttpLogRecordExporter);
+    Mockito.when(mockOtlpHttpLogRecordExporter.shutdown())
+        .thenReturn(CompletableResultCode.ofSuccess());
+    Mockito.lenient()
+        .when(mockOtlpHttpLogRecordExporter.flush())
+        .thenReturn(CompletableResultCode.ofSuccess());
+    Mockito.when(mockOtlpHttpLogRecordExporter.toBuilder())
+        .thenReturn(spyOtlpHttpLogRecordExporterBuilder);
+    Mockito.when(mockOtlpHttpLogRecordExporter.export(Mockito.anyCollection()))
+        .thenAnswer(
+            invocationOnMock -> {
+              exportedLogRecordContainer.addAll(invocationOnMock.getArgument(0));
+              return CompletableResultCode.ofSuccess();
+            });
+  }
+
+  // Configure necessary behavior on the gRPC mock log record exporters to work.
+  // Mockito.lenient is used here because this method is used with parameterized tests where based
+  // on certain inputs, certain stubbings may not be required.
+  private static void configureGrpcMockLogRecordExporter(
+      OtlpGrpcLogRecordExporter mockOtlpGrpcLogRecordExporter,
+      OtlpGrpcLogRecordExporterBuilder spyOtlpGrpcLogRecordExporterBuilder,
+      List<LogRecordData> exportedLogRecordContainer) {
+    Mockito.lenient()
+        .when(spyOtlpGrpcLogRecordExporterBuilder.build())
+        .thenReturn(mockOtlpGrpcLogRecordExporter);
+    Mockito.lenient()
+        .when(mockOtlpGrpcLogRecordExporter.shutdown())
+        .thenReturn(CompletableResultCode.ofSuccess());
+    Mockito.lenient()
+        .when(mockOtlpGrpcLogRecordExporter.flush())
+        .thenReturn(CompletableResultCode.ofSuccess());
+    Mockito.lenient()
+        .when(mockOtlpGrpcLogRecordExporter.toBuilder())
+        .thenReturn(spyOtlpGrpcLogRecordExporterBuilder);
+    Mockito.lenient()
+        .when(mockOtlpGrpcLogRecordExporter.export(Mockito.anyCollection()))
+        .thenAnswer(
+            invocationOnMock -> {
+              exportedLogRecordContainer.addAll(invocationOnMock.getArgument(0));
+              return CompletableResultCode.ofSuccess();
+            });
+  }
+
   @AutoValue
   abstract static class ProjectIdTestBehavior {
     // A null user specified project ID represents the use case where user omits specifying it
@@ -1148,6 +1444,8 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
 
     abstract boolean getExpectedIsMetricsSignalModified();
 
+    abstract boolean getExpectedIsLogsSignalModified();
+
     static Builder builder() {
       return new AutoValue_GcpAuthAutoConfigurationCustomizerProviderTest_TargetSignalBehavior
           .Builder();
@@ -1166,6 +1464,10 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
       // Set whether the combination of specified OTel properties and configured target signals
       // should lead to modification of the OTLP metrics exporters.
       abstract Builder setExpectedIsMetricsSignalModified(boolean expectedModified);
+
+      // Set whether the combination of specified OTel properties and configured target signals
+      // should lead to modification of the OTLP log record exporters.
+      abstract Builder setExpectedIsLogsSignalModified(boolean expectedModified);
 
       abstract TargetSignalBehavior build();
     }
@@ -1328,31 +1630,62 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
 
   private OpenTelemetrySdk buildOpenTelemetrySdkWithExporter(SpanExporter spanExporter) {
     return buildOpenTelemetrySdkWithExporter(
-        spanExporter, OtlpHttpMetricExporter.getDefault(), defaultOtelPropertiesSpanExporter);
+        spanExporter,
+        OtlpHttpMetricExporter.getDefault(),
+        OtlpHttpLogRecordExporter.getDefault(),
+        defaultOtelPropertiesSpanExporter);
   }
 
   @SuppressWarnings("UnusedMethod")
   private OpenTelemetrySdk buildOpenTelemetrySdkWithExporter(
       SpanExporter spanExporter, ImmutableMap<String, String> customOTelProperties) {
     return buildOpenTelemetrySdkWithExporter(
-        spanExporter, OtlpHttpMetricExporter.getDefault(), customOTelProperties);
+        spanExporter,
+        OtlpHttpMetricExporter.getDefault(),
+        OtlpHttpLogRecordExporter.getDefault(),
+        customOTelProperties);
   }
 
   private OpenTelemetrySdk buildOpenTelemetrySdkWithExporter(MetricExporter metricExporter) {
     return buildOpenTelemetrySdkWithExporter(
-        OtlpHttpSpanExporter.getDefault(), metricExporter, defaultOtelPropertiesMetricExporter);
+        OtlpHttpSpanExporter.getDefault(),
+        metricExporter,
+        OtlpHttpLogRecordExporter.getDefault(),
+        defaultOtelPropertiesMetricExporter);
   }
 
   @SuppressWarnings("UnusedMethod")
   private OpenTelemetrySdk buildOpenTelemetrySdkWithExporter(
       MetricExporter metricExporter, ImmutableMap<String, String> customOtelProperties) {
     return buildOpenTelemetrySdkWithExporter(
-        OtlpHttpSpanExporter.getDefault(), metricExporter, customOtelProperties);
+        OtlpHttpSpanExporter.getDefault(),
+        metricExporter,
+        OtlpHttpLogRecordExporter.getDefault(),
+        customOtelProperties);
+  }
+
+  private OpenTelemetrySdk buildOpenTelemetrySdkWithExporter(LogRecordExporter logRecordExporter) {
+    return buildOpenTelemetrySdkWithExporter(
+        OtlpHttpSpanExporter.getDefault(),
+        OtlpHttpMetricExporter.getDefault(),
+        logRecordExporter,
+        defaultOtelPropertiesLogRecordExporter);
+  }
+
+  @SuppressWarnings("UnusedMethod")
+  private OpenTelemetrySdk buildOpenTelemetrySdkWithExporter(
+      LogRecordExporter logRecordExporter, ImmutableMap<String, String> customOtelProperties) {
+    return buildOpenTelemetrySdkWithExporter(
+        OtlpHttpSpanExporter.getDefault(),
+        OtlpHttpMetricExporter.getDefault(),
+        logRecordExporter,
+        customOtelProperties);
   }
 
   private OpenTelemetrySdk buildOpenTelemetrySdkWithExporter(
       SpanExporter spanExporter,
       MetricExporter metricExporter,
+      LogRecordExporter logRecordExporter,
       ImmutableMap<String, String> customOtelProperties) {
     SpiHelper spiHelper =
         SpiHelper.create(GcpAuthAutoConfigurationCustomizerProviderTest.class.getClassLoader());
@@ -1387,6 +1720,22 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
                                 public MetricExporter createExporter(
                                     ConfigProperties configProperties) {
                                   return metricExporter;
+                                }
+
+                                @Override
+                                public String getName() {
+                                  return "otlp";
+                                }
+                              }));
+                    }
+                    if (spiClass == ConfigurableLogRecordExporterProvider.class) {
+                      return Collections.singletonList(
+                          spiClass.cast(
+                              new ConfigurableLogRecordExporterProvider() {
+                                @Override
+                                public LogRecordExporter createExporter(
+                                    ConfigProperties configProperties) {
+                                  return logRecordExporter;
                                 }
 
                                 @Override
@@ -1432,6 +1781,17 @@ class GcpAuthAutoConfigurationCustomizerProviderTest {
     long workOutput = busyloop();
     long randomValue = TEST_RANDOM.nextInt(1000);
     longCounter.add(randomValue, Attributes.of(AttributeKey.longKey("work_loop"), workOutput));
+  }
+
+  private static void generateTestLogRecord(OpenTelemetrySdk openTelemetrySdk) {
+    long workOutput = busyloop();
+    openTelemetrySdk
+        .getSdkLoggerProvider()
+        .get("test")
+        .logRecordBuilder()
+        .setBody("sample log record")
+        .setAllAttributes(Attributes.of(AttributeKey.longKey("work_loop"), workOutput))
+        .emit();
   }
 
   // loop to simulate work done